feat(example): add SceneManager
