+++ packages/RNExternalDisplayFabricExample/index.js
@@ -4,7 +4,7 @@
 
 import { AppRegistry } from 'react-native'
 
-import App from 'rn-external-display-examples/js/App'
+import App from 'rn-external-display-examples/App'
 import { name as appName } from './app.json'
 
 AppRegistry.registerComponent(appName, () => App)
