+++ packages/rnexternal-display-examples/SceneManager.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import { View, Button } from 'react-native'
+import ExternalDisplayEvent from 'react-native-external-display/js/NativeRNExternalDisplayEvent'
+
+export default function SceneManager(props) {
+  return (
+    <>
+      <Button
+        title="REQUEST NEW SCENE"
+        onPress={() => {
+          ExternalDisplayEvent.requestScene().then(console.log)
+        }}
+      />
+    </>
+  )
+}
