+++ packages/rnexternal-display-examples/SimpleTextInterval.js
@@ -1,9 +1,9 @@
 // @flow
 
-import React, { useState } from 'react'
-import { SafeAreaView, View, Button } from 'react-native'
-import WebView from 'react-native-webview'
+import React, { useEffect, useState } from 'react'
+import { SafeAreaView, Text, View, Button } from 'react-native'
 import ExternalDisplay, { getScreens } from 'react-native-external-display'
+import SceneManager from './SceneManager'
 
 type Props = {
   onBack: () => void,
@@ -11,9 +11,15 @@ type Props = {
 
 export default function Example(props: Props) {
   const { onBack } = props
+  const [t, setT] = useState(0)
   const [info, setInfo] = useState(getScreens())
+
   const [on, setOn] = useState(true)
   const [mount, setMount] = useState(true)
+  useEffect(() => {
+    const interval = setInterval(() => setT((d) => d + 1), 1000)
+    return () => clearInterval(interval)
+  }, [])
   return (
     <SafeAreaView
       style={{
@@ -32,12 +38,17 @@ export default function Example(props: Props) {
             onScreenConnect={setInfo}
             onScreenDisconnect={setInfo}
           >
-            <WebView
-              source={{ uri: 'https://reactnative.dev/' }}
-              allowsInlineMediaPlayback
-              allowsFullscreenVideo
-              style={{ flex: 1 }}
-            />
+            <View
+              style={{
+                flex: 1,
+                justifyContent: 'center',
+                alignItems: 'center',
+                backgroundColor: '#333',
+                height: 100,
+              }}
+            >
+              <Text style={{ color: 'red', fontSize: 40 }}>{t}</Text>
+            </View>
           </ExternalDisplay>
         )}
       </View>
@@ -46,6 +57,7 @@ export default function Example(props: Props) {
         onPress={() => setMount((d) => !d)}
         title={mount ? 'UNMOUNT' : 'MOUNT'}
       />
+      <SceneManager />
       <Button onPress={onBack} title="BACK" />
     </SafeAreaView>
   )
